+++ src/php/Console/ConsoleFactory.php
@@ -5,11 +5,17 @@ declare(strict_types=1);
 namespace Phel\Console;
 
 use Gacela\Framework\AbstractFactory;
+use Phel\Filesystem\FilesystemFacadeInterface;
 
 final class ConsoleFactory extends AbstractFactory
 {
     public function getConsoleCommands(): array
     {
         return $this->getProvidedDependency(ConsoleDependencyProvider::COMMANDS);
+    }
+
+    public function getFilesystemFacade(): FilesystemFacadeInterface
+    {
+        return $this->getProvidedDependency(ConsoleDependencyProvider::FACADE_FILESYSTEM);
     }
 }