+++ src/php/Filesystem/FilesystemFacadeInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Phel\Filesystem;
+
+interface FilesystemFacadeInterface
+{
+    public function addFile(string $file): void;
+
+    public function clearAll(): void;
+}